requirements.txt
# Trading System Full - Requirements
# Install with: pip install -r requirements.txt

# Data
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0

# Database
# sqlite3 is built-in

# Configuration
pyyaml>=6.0

# Dashboard
streamlit>=1.30.0
plotly>=5.18.0

# Progress bars
tqdm>=4.66.0

# Date handling
python-dateutil>=2.8.2
